+++ src/component/StyledComponents.ts
@@ -16,13 +16,17 @@ import {
   PolyTouchableWithoutFeedback,
   PolyView,
 }                              from "./PolyComponents";
-import {classesId}             from "..";
+import {classesId}      from "..";
 import {PropsOf, shallowEqual} from "../Utils";
 import Styler, {StylableProps} from "./Styler";
 
-export type StyledComponent<P> = React.ComponentType<Omit<P, "style"> & StylableProps>;
+export type StyledComponent<P> = React.ComponentType<(keyof P extends "style" ? Omit<P, "style">: P) & StylableProps>;
 
-export const styled = <P>(baseComponent: React.ComponentType<P>): StyledComponent<P> => {
+export interface StyledOptions {
+  autoFlattens?: boolean;
+}
+
+export const styled = <P>(baseComponent: React.ComponentType<P>, options?: StyledOptions): StyledComponent<P> => {
   return React.memo(Object.assign(React.forwardRef((props: any, ref) => {
     const {children, style, pseudoClasses, classes, ...otherProps} = props;
 
@@ -33,6 +37,7 @@ export const styled = <P>(baseComponent: React.ComponentType<P>): StyledComponen
       ref,
       _baseComponent: baseComponent,
       children      : React.createElement(baseComponent, otherProps as any, children),
+      options,
     } as any);
   }), {displayName: `Styler${baseComponent.displayName ? `[${baseComponent.displayName}]` : ""}`}), styledArePropsEqual);
 };
@@ -44,17 +49,17 @@ export function styledArePropsEqual<T>(prevProps: T, nextProps: T): boolean {
 }
 
 // we have to be explict about the type of the decl here because typescript expands the keyof otherwise https://github.com/microsoft/TypeScript/issues/27171
-export const StyledSafeAreaView             :StyledComponent<PropsOf<typeof SafeAreaView>> = styled(SafeAreaView);
-export const StyledView                     :StyledComponent<PropsOf<typeof PolyView>> = styled(PolyView);
-export const StyledText                     :StyledComponent<PropsOf<typeof PolyText>> = styled(PolyText);
-export const StyledTextInput                :StyledComponent<PropsOf<typeof TextInput>> = styled(TextInput);
-export const StyledButton                   :StyledComponent<PropsOf<typeof Button>> = styled(Button);
-export const StyledTouchableNativeFeedback  :StyledComponent<PropsOf<typeof PolyTouchableNativeFeedback>> = styled(PolyTouchableNativeFeedback);
-export const StyledTouchableOpacity         :StyledComponent<PropsOf<typeof PolyTouchableOpacity>> = styled(PolyTouchableOpacity);
-export const StyledTouchableWithoutFeedback :StyledComponent<PropsOf<typeof PolyTouchableWithoutFeedback>> = styled(PolyTouchableWithoutFeedback);
-export const StyledTouchableHighlight       :StyledComponent<PropsOf<typeof PolyTouchableHighlight>> = styled(PolyTouchableHighlight);
-export const StyledScrollView               :StyledComponent<PropsOf<typeof ScrollView>> = styled(ScrollView);
-export const StyledImage                    :StyledComponent<PropsOf<typeof Image>> = styled(Image);
+export const StyledSafeAreaView             : StyledComponent<PropsOf<typeof SafeAreaView>> = styled(SafeAreaView, {autoFlattens: true});
+export const StyledView                     : StyledComponent<PropsOf<typeof PolyView>> = styled(PolyView, {autoFlattens: true});
+export const StyledText                     : StyledComponent<PropsOf<typeof PolyText>> = styled(PolyText, {autoFlattens: true});
+export const StyledTextInput                : StyledComponent<PropsOf<typeof TextInput>> = styled(TextInput, {autoFlattens: true});
+export const StyledButton                   : StyledComponent<PropsOf<typeof Button>> = styled(Button, {autoFlattens: true});
+export const StyledTouchableNativeFeedback  : StyledComponent<PropsOf<typeof PolyTouchableNativeFeedback>> = styled(PolyTouchableNativeFeedback, {autoFlattens: true});
+export const StyledTouchableOpacity         : StyledComponent<PropsOf<typeof PolyTouchableOpacity>> = styled(PolyTouchableOpacity, {autoFlattens: true});
+export const StyledTouchableWithoutFeedback : StyledComponent<PropsOf<typeof PolyTouchableWithoutFeedback>> = styled(PolyTouchableWithoutFeedback, {autoFlattens: true});
+export const StyledTouchableHighlight       : StyledComponent<PropsOf<typeof PolyTouchableHighlight>> = styled(PolyTouchableHighlight, {autoFlattens: true});
+export const StyledScrollView               : StyledComponent<PropsOf<typeof ScrollView>> = styled(ScrollView, {autoFlattens: true});
+export const StyledImage                    : StyledComponent<PropsOf<typeof Image>> = styled(Image, {autoFlattens: true});
 
 export type StyledSafeAreaViewProps = PropsOf<typeof StyledSafeAreaView>;
 export type StyledViewProps = PropsOf<typeof StyledView>;
@@ -69,8 +74,8 @@ export type StyledScrollViewProps = PropsOf<typeof StyledScrollView>;
 export type StyledImageProps = PropsOf<typeof StyledImage>;
 
 export const StyledAnimated = {
-  View      : styled(Animated.View),
-  Text      : styled(Animated.Text),
-  ScrollView: styled(Animated.ScrollView),
-  Image     : styled(Animated.Image),
+  View      : styled(Animated.View, {autoFlattens: true}),
+  Text      : styled(Animated.Text, {autoFlattens: true}),
+  ScrollView: styled(Animated.ScrollView, {autoFlattens: true}),
+  Image     : styled(Animated.Image, {autoFlattens: true}),
 };